requirements.txt
pytube
numpy
youtube-search-python
python-slugify
shutil